rename app folder to api
