+++ api/src/exceptions.py
