+++ api/src/utils.py
@@ -1,7 +1,7 @@
 from enum import Enum
 
-from app.src import constants, exceptions
-from app.src.initializer import logger
+from api.src import constants, exceptions
+from api.src.initializer import logger
 
 
 def is_leap_year(year: int) -> bool:
