+++ api/tests/src/utils_test.py
@@ -1,8 +1,8 @@
 import unittest
 
-from app.src.constants import MONTH, PIVOT_DATE, Date
-from app.src.exceptions import InvalidDateFormat
-from app.src.utils import (
+from api.src.constants import MONTH, PIVOT_DATE, Date
+from api.src.exceptions import InvalidDateFormat
+from api.src.utils import (
     count_leap_years,
     date_validator,
     get_actual_days_in_month,
